Move FileMetadataAccessor to :files
