+++ platforms/core-runtime/files/src/main/java/org/gradle/internal/file/FileMetadataAccessor.java
@@ -14,9 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.nativeintegration.filesystem;
-
-import org.gradle.internal.file.FileMetadata;
+package org.gradle.internal.file;
 
 import java.io.File;
