+++ platforms/core-runtime/native/src/main/java/org/gradle/internal/nativeintegration/filesystem/services/GenericFileSystem.java
@@ -22,8 +22,8 @@ import org.gradle.api.internal.file.temp.TemporaryFileProvider;
 import org.gradle.internal.SystemProperties;
 import org.gradle.internal.file.FileException;
 import org.gradle.internal.file.FileMetadata;
+import org.gradle.internal.file.FileMetadataAccessor;
 import org.gradle.internal.file.StatStatistics;
-import org.gradle.internal.nativeintegration.filesystem.FileMetadataAccessor;
 import org.gradle.internal.nativeintegration.filesystem.FileModeAccessor;
 import org.gradle.internal.nativeintegration.filesystem.FileModeMutator;
 import org.gradle.internal.nativeintegration.filesystem.FileSystem;
