+++ platforms/core-runtime/native/src/test/groovy/org/gradle/internal/nativeintegration/filesystem/services/AbstractFileMetadataAccessorTest.groovy
@@ -19,8 +19,8 @@ package org.gradle.internal.nativeintegration.filesystem.services
 import org.gradle.api.UncheckedIOException
 import org.gradle.internal.file.FileMetadata
 import org.gradle.internal.file.FileMetadata.AccessType
+import org.gradle.internal.file.FileMetadataAccessor
 import org.gradle.internal.file.FileType
-import org.gradle.internal.nativeintegration.filesystem.FileMetadataAccessor
 import org.gradle.test.fixtures.file.TestNameTestDirectoryProvider
 import org.gradle.test.precondition.Requires
 import org.gradle.test.preconditions.UnitTestPreconditions
