+++ platforms/core-runtime/native/src/test/groovy/org/gradle/internal/nativeintegration/filesystem/services/FallbackFileMetadataAccessorTest.groovy
@@ -17,7 +17,7 @@
 package org.gradle.internal.nativeintegration.filesystem.services
 
 import org.gradle.internal.file.FileMetadata
-import org.gradle.internal.nativeintegration.filesystem.FileMetadataAccessor
+import org.gradle.internal.file.FileMetadataAccessor
 import org.gradle.util.UsesNativeServices
 
 import static org.gradle.internal.file.FileMetadata.AccessType.DIRECT
